[fix/aut826] moved ParametersTableWidget to its own File.
