+++ src/smartpeak/include/SmartPeak/ui/sources.cmake
@@ -8,6 +8,7 @@ set(sources_list_h
 	FilePicker.h
 	GuiAppender.h
 	Heatmap2DWidget.h
+	ParametersTableWidget.h
 	Report.h
 	SpectraPlotWidget.h
 	Widget.h
